+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.2.5 (25/09/2024)
 
 * Implement ESM embedding encoder ([#33](https://github.com/a-r-j/ProteinWorkshop/pull/33), [#41](https://github.com/a-r-j/ProteinWorkshop/pull/33))
+* Add script for plotting UMAP embeddings of any dataset given a pre-trained encoder model
 
 ### 0.2.4 (10/09/2024)